Moved non-presenter specific validators from 'br/presenter/validator' to 'br/validation'.
