+++ brjs-sdk/sdk/libs/javascript/br-presenter/src/br/presenter/validator/.js-style
@@ -1,0 +1,1 @@
+common-js
