+++ brjs-sdk/sdk/libs/javascript/br-presenter/src/br/presenter/validator/CrossValidationPropertyListener.js
@@ -2,25 +2,28 @@
  * @module br/presenter/validator/CrossValidationPropertyListener
  */
 
+var PropertyListener = require('br/presenter/property/PropertyListener');
+
 /**
  * @private
  * @class
  * @alias module:br/presenter/validator/CrossValidationPropertyListener
  * @implements module:br/presenter/property/PropertyListener
- * 
+ *
  * @param oCrossPropertyValidationProxy
  * @param sPropertyId
  */
-br.presenter.validator.CrossValidationPropertyListener = function(oCrossPropertyValidationProxy, sPropertyId)
+function CrossValidationPropertyListener(oCrossPropertyValidationProxy, sPropertyId)
 {
 	this.m_oValidationProxy = oCrossPropertyValidationProxy;
 	this.m_sPropertyId = sPropertyId;
-};
+}
 
-br.Core.inherit(br.presenter.validator.CrossValidationPropertyListener, br.presenter.property.PropertyListener);
+br.Core.inherit(CrossValidationPropertyListener, PropertyListener);
 
-br.presenter.validator.CrossValidationPropertyListener.prototype.onPropertyChanged = function()
+CrossValidationPropertyListener.prototype.onPropertyChanged = function()
 {
 	this.m_oValidationProxy._$onPropertyChanged(this.m_sPropertyId);
 };
 
+module.exports = CrossValidationPropertyListener;
