+++ brjs-sdk/sdk/libs/javascript/br-presenter/src/br/presenter/validator/NotEmptyValidator.js
@@ -1,32 +1,1 @@
-/**
- * @module br/presenter/validator/NotEmptyValidator
- */
-
-/**
- * @class
- * @alias module:br/presenter/validator/NotEmptyValidator
- * @implements module:br/presenter/validator/Validator
- */
-br.presenter.validator.NotEmptyValidator = function(sFailureMessage)
-{
-	this.sMessage = sFailureMessage;
-};
-
-br.Core.implement(br.presenter.validator.NotEmptyValidator, br.presenter.validator.Validator);
-
-br.presenter.validator.NotEmptyValidator.prototype.validate = function(vValue, mAttributes, oValidationResult)
-{
-	if(vValue=="")
-	{
-		var bIsValid = false;
-	}
-	else
-	{
-		var bIsValid = true;
-	}
-	
-	var oTranslator = require("br/I18n").getTranslator();
-	var sFailureMessage = oTranslator.tokenExists(this.sMessage) ? oTranslator.getMessage(this.sMessage) : this.sMessage;
-	
-	oValidationResult.setResult(bIsValid, sFailureMessage);
-};
+module.exports = require('br/validation/NotEmptyValidator');
