+++ brjs-sdk/sdk/libs/javascript/br-presenter/src/br/presenter/validator/ValidSelectionValidator.js
@@ -2,31 +2,34 @@
  * @module br/presenter/validator/ValidSelectionValidator
  */
 
+var Validator = require('br/validation/Validator');
+var Errors = require('br/Errors');
+
 /**
  * @private
  * @class
  * @alias module:br/presenter/validator/ValidSelectionValidator
- * @implements module:br/presenter/validator/Validator
- * 
+ * @implements module:br/validation/Validator
+ *
  * @param {module:br/presenter/node/OptionsNodeList} oOptions The list of valid options.
  */
-br.presenter.validator.ValidSelectionValidator = function(oOptions)
+function ValidSelectionValidator(oOptions)
 {
 	if (!oOptions || !(oOptions instanceof br.presenter.node.OptionsNodeList))
 	{
-		throw new br.Errors.InvalidParametersError("You must provide an instance of OptionsNodeList");
+		throw new Errors.InvalidParametersError("You must provide an instance of OptionsNodeList");
 	}
 	this.m_oOptions = oOptions;
 	this.m_bAllowInvalidSelections = false;
-};
-br.Core.implement(br.presenter.validator.ValidSelectionValidator, br.presenter.validator.Validator);
+}
+br.Core.implement(ValidSelectionValidator, Validator);
 
-br.presenter.validator.ValidSelectionValidator.prototype.allowInvalidSelections = function(bAllowInvalidSelections)
+ValidSelectionValidator.prototype.allowInvalidSelections = function(bAllowInvalidSelections)
 {
 	this.m_bAllowInvalidSelections = bAllowInvalidSelections;
 };
 
-br.presenter.validator.ValidSelectionValidator.prototype.validate = function(vValue, mAttributes, oValidationResult)
+ValidSelectionValidator.prototype.validate = function(vValue, mAttributes, oValidationResult)
 {
 	// default results, unless value is found in the selection options
 	var bIsValid = false;
@@ -56,3 +59,4 @@ br.presenter.validator.ValidSelectionValidator.prototype.validate = function(vVa
 	oValidationResult.setResult(bIsValid, sValidationMessage);
 };
 
+module.exports = ValidSelectionValidator;
