--- brjs-sdk/sdk/libs/javascript/br-presenter/test-unit/resources/en.properties
@@ -1,1 +1,0 @@
-br.presenter.i18ntesttoken=i18nErrorMessage
