+++ brjs-sdk/sdk/libs/javascript/br-validation/.js-style
@@ -1,0 +1,1 @@
+namespaced-js
