+++ brjs-sdk/sdk/libs/javascript/br-validation/br-lib.conf
@@ -1,0 +1,1 @@
+requirePrefix: br/validation
