+++ brjs-sdk/sdk/libs/javascript/br-validation/test-unit/jsTestDriver.conf
@@ -1,0 +1,13 @@
+server: http://localhost:4224
+basepath: .
+
+load:
+  - bundles/i18n/i18n.bundle
+  - bundles/js/js.bundle
+
+serve:
+  - bundles/xml.bundle
+  - bundles/html.bundle
+
+test:
+  - tests/**.js
