+++ brjs-sdk/sdk/libs/javascript/br-validation/test-unit/resources/en.properties
@@ -1,0 +1,1 @@
+br.validation.i18ntesttoken=i18nErrorMessage
